% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.oglmx}
\alias{AIC.oglmx}
\title{Akaike's An Information Criterion for \code{oglm} objects}
\usage{
\method{AIC}{oglmx}(object, ..., k = 2)
}
\arguments{
\item{object}{a fitted model object for which there exists a
    \code{logLik} method to extract the corresponding log-likelihood, or
    an object inheriting from class \code{logLik}.}

\item{...}{optionally more fitted model objects.}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\description{
Compute \code{AIC} for ordered and interval
regression objects. See \code{\link[stats:AIC]{stats::AIC()}} for the rest
}
\details{
See \code{\link[stats:AIC]{stats::AIC()}}
}
\seealso{
\code{\link[stats:AIC]{stats::AIC()}},
\code{\link[stats:logLik]{stats::logLik()}}, \code{\link[stats:nobs]{stats::nobs()}}
}
