% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{formula.oglmx}
\alias{formula.oglmx}
\title{Obtain model formula for an \code{oglm} object}
\usage{
\method{formula}{oglmx}(x, ...)
}
\arguments{
\item{x}{object of class \code{oglmx}}

\item{...}{Additional arguments, currently ignored.}
}
\value{
an object of class \code{formula}.
}
\description{
Given an object of class \code{oglm} the function describes
the estimated model via an expression of class \code{\link[stats:formula]{stats::formula()}}.
The function serves to provide a name of a model to the \code{\link[lmtest:lrtest]{lmtest::lrtest()}}
}
