% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_distrib.R
\name{pgumbel}
\alias{pgumbel}
\alias{dgumbel}
\alias{pGumbel}
\alias{dGumbel}
\alias{rgumbel}
\alias{rGumbel}
\title{Gumbel distribution functions}
\usage{
pgumbel(q, loc = 0, scale = 1, lower.tail = TRUE)

dgumbel(x, loc = 0, scale = 1, log = FALSE)

pGumbel(q, loc = 0, scale = 1, lower.tail = TRUE)

dGumbel(x, loc = 0, scale = 1, log = FALSE)

rgumbel(n, loc = 0, scale = 1)

rGumbel(n, loc = 0, scale = 1)
}
\arguments{
\item{q}{vector of quantiles.}

\item{loc, scale}{Location and scale parameters.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{x}{vector of quantiles.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}
}
\value{
\code{dgumbel} gives the density, \code{pgumbel} gives the distribution function,
and \code{rgumbel} generates random deviates.
}
\description{
Density, distribution function, quantile function
and random generation for the gumbel distribtion. Extracted
from \link[MASS]{polr} source code
}
