% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.oglmx}
\alias{predict.oglmx}
\title{Predictions from ordered discrete model}
\usage{
\method{predict}{oglmx}(
  object,
  newdata = NULL,
  type = c("class", "probs", "latent", "xb"),
  ...
)
}
\arguments{
\item{object}{An ordered discrete model computed using \code{oglmx} package}

\item{newdata}{optionally, a data frame in which to look for variables with
    which to predict.  If omitted, the fitted linear predictors are used.}

\item{type}{Prediction considered. Can be \emph{'class'}
(predicted label as maximum probability label, default);
\emph{'probs'} (probabilities of each observation to belong to a particular class)
or \emph{'latent'} (\eqn{Y^*} latent variable)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Depends of the value of prediction \code{type} \describe{
\item{class}{Most likely label, i.e. \eqn{l = \arg \max_l p_j}}
\item{probs}{Prediced probabilities for each label, i.e. \eqn{p_ij} matrix with
$i$ observation index and $j$ class label}
\item{latent}{Prediced value in latent space \eqn{y^*}}
}
}
\description{
Predictions from ordered discrete model
}
